Resubmit to add message
Finished up a few select statements relevant to information extraction for different UI elements and other crap. I added some commented notes on the bottom for personal reference (since I will be adding a few more queries) and also for you guys to look at and read over to make sure we are on track.